stream get and put through pipes instead of the bus
the bus has a hard time handling the amounts of data we may conceivably
be pushing over it (gibibytes... literally). to mitigate enormous
performance problems push the data through a pipe instead and only pass
the unix socket over the bus. this is a bit more code and a bit crappy
posix-y one at that, but it's heaps better than hogging the system bus
and perhaps even get stuck because of problems in qdbus with large
dbusmessages.

CCBUG: 468673