Add method to notify about expired authorization
Previous to this commit when the authorization expired this would
happen completely silently without any way to notice this
programmatically. The surprise only came when the user tried to do
any action, which immediately would show a password prompt. There
wasn't really a convenient way to work around that.

This commit introduces a special method which will only return once
the authorization has expired.

Required for:
https://invent.kde.org/system/dolphin/-/merge_requests/784